cm.backend/requirements.txt
annotated-types==0.7.0

anyio==4.8.0

bcrypt==4.2.1

cffi==1.17.1

click==8.1.8

colorama==0.4.6

cryptography==44.0.1

fastapi==0.115.8

h11==0.14.0

idna==3.10

iniconfig==2.0.0

packaging==24.2

passlib==1.7.4

pluggy==1.5.0

pycparser==2.22

pycryptodome==3.21.0

pydantic==2.10.6

pydantic_core==2.27.2

PyJWT==2.10.1

pytest==8.3.4

python-multipart==0.0.20

setuptools==75.8.0

sniffio==1.3.1

starlette==0.45.3

typing_extensions==4.12.2

uvicorn==0.34.0

wheel==0.45.1